Praktikum 3

+++ Praktikum 3/kalkulator/main.py
@@ -8,8 +8,8 @@ def konfirmasi():
 print("=== KALKULATOR ===")
 while True:
     try:
-        a = int(input("Masukkan angka pertama: "))
-        b = int(input("Masukkan angka kedua: "))
+        a = float(input("Masukkan angka pertama: "))
+        b = float(input("Masukkan angka kedua: "))
     except ValueError:
         print("Error: Input harus berupa angka")
         konfirmasi()
@@ -18,7 +18,7 @@ while True:
         print("2. Kurang")
         print("3. Kali")
         print("4. Bagi")
-        pilihan = float(input("Masukkan pilihan: "))
+        pilihan = int(input("Masukkan pilihan: "))
         if pilihan == 1:
             print("Hasil: ", kalkulator.tambah(a, b))
             konfirmasi()
